#3 - Refactor popup code to make it more DRY
