+++ src/html/popup.html
@@ -7,6 +7,7 @@
   </head>
   <body>
     <p>To quit, press <kbd class="keyboard-key">esc</kbd></p>
+    <p>To refresh extension, press <kbd class="keyboard-key">R</kbd></p>
     <script src="../scripts/popup.js" charset="utf-8"></script>
   </body>
 </html>
